Version 2.1.0 1. Support customize numbers and positions of pieces and boom-buttons. 2. Support draggable. 3. Support fade-views on BMB. 4. Support 3D animations of boom-buttons. 5. Correct some orders of piece-placing and button-placing. Now pieces and boom-buttons are placed from left-top to right-bottom. 6. Support auto-boom and auto-boom-immediately of BMB, which is helpful when BMB is supposed to boom at start-up.

+++ boommenu/src/main/java/com/nightonke/boommenu/Piece/Dot.java
@@ -21,7 +21,7 @@ final class Dot extends BoomPiece {
     }
 
     @Override
-    public void init(int color, int cornerRadius) {
+    public void init(int color, float cornerRadius) {
         Drawable backgroundDrawable;
         if (cornerRadius < 0) backgroundDrawable = Util.getDrawable(this, R.drawable.piece_dot, null)
                 .mutate();
